+++ Agencias.Api.Domain/Dtos/CategoriaDto.cs
@@ -28,7 +28,7 @@ namespace Agencias.Api.Domain.Dtos
 		public int Id { get; set; }
 		public int Idcategoria { get; set; }
 		public int Antiguedad { get; set; }
-		public int Tope_Ventas { get; set; }
+		public Double Tope_Ventas { get; set; }
 		public Double Por_ProdPro { get; set; }
 		public Double Por_ProdSup { get; set; }
 		public Double Por_ProdAge { get; set; }
